// src/app/models/put-empleado.interface.ts
export interface PutEmpleado {
  idEmpleado: number;
  primerNombre: string;
  segundoNombre?: string;
  primerApellido: string;
  segundoApellido: string;
  idTipoIdent: number;
  numeroIdentificacion: string;
  fechaNacimiento: Date;
  fechaContratacion: Date;
  fechaTerminacion: Date;
  emailPrincipal: string;
  telefonoFijo: number;
  telefonoCelular: number;
  rol: string;
}